feature:add-token-post

+++ src/resolvers/Mutation.js
@@ -1,6 +1,8 @@
 import bcrypt from 'bcryptjs'
 import jwt from 'jsonwebtoken'
+import getUserId from '../utils/getUserId'
 //Taken in password -> Validate
+
 
 const Mutation = {
     async createUser(parent, args, { prisma }, info) {
@@ -17,23 +19,45 @@ const Mutation = {
             token: jwt.sign({userId:user.id},'thisisasecret')
         }
     },
-    async deleteUser(parent, args, { prisma }, info) {
+    async login(parent,args,{prisma},info) {
+        const user = await prisma.query.user({
+            where: {
+                email:args.data.email
+            }
+        })
+        if(!user){
+            throw new Error('Unabled to login')
+        }
+        const isMatch = await bcrypt.compare(args.data.password, user.password);
+        if(!isMatch){
+            throw new Error('Unabled to login')
+        }
+        return {
+            user,
+            token: jwt.sign({userId:user.id},'thisisasecret')
+        }
+    },
+    async deleteUser(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
         return prisma.mutation.deleteUser({ 
             where: { 
-                id:args.id 
+                id:userId
             } 
         },info)
     },
-    async updateUser(parent, args, { prisma }, info) {
+    async updateUser(parent, args, { prisma,request }, info) {
+        const userId = getUserId(request)
         return prisma.mutation.updateUser({
             where: {
-                id:args.id
+                id:userId
             },
             data:args.data
         },info)
 
     },
-    createPost(parent, args, { prisma }, info) {
+    createPost(parent, args, { prisma,request }, info) {
+        const userId = getUserId(request)
+
         return prisma.mutation.createPost({
             data: {
                 title:args.data.title,
@@ -41,17 +65,27 @@ const Mutation = {
                 published:args.data.published,
                 author:{
                     connect:{
-                        id: args.data.author
+                        id: userId
                     }
                 }
             }
         },info)
 
     },
-    deletePost(parent, args, { prisma }, info) {
+    async deletePost(parent, args, { prisma,request }, info) {
+        const userId = getUserId(request)
+        const postExists = await prisma.exists.Post({
+            id:args.id,
+            author:{
+                id:userId
+            }
+        })
+        if(!postExists){
+            throw new Error('unable to delete post')
+        }
         return prisma.mutation.deletePost({
             where: {
-                id:args.id
+                id:args.id,
             }
         },info)
     },
